requirements.txt
alembic==0.9.7
click==6.7
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
psycopg2==2.7.3.2
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.2
SQLAlchemy-ImageAttach==1.1.0
visitor==0.1.3
Wand==0.4.4
Werkzeug==0.14.1
WTForms==2.1
